schema_registry: take subject by value in sharded_store
Make the API of the sharded_store consistently take all parameters by
value instead of by reference.

This method was the only outlier. We have to copy the value of the
subject into the lambda anyway, so this only introduces an extra move.
However, it makes the async API of sharded_store easier to reason about.