lsm: add tests for compaction pointers
The compaction pointer is important to ensure we rotate properly through
the entire key space when computing compactions. Without it, we could
starve part of the key space from compactions.

We just add tests for this and restore a helpful comment from the
leveldb source about why we immediately update the compaction pointer in
addition to applying it to the edit.