rpk: validate secret existence before creation.
Secret exists in the requested shadow cluster, in
rpk if we are logged in to the shadow cluster we
can validate this upfront.

The create request goes to the dataplane which
doesn't have access to the secret store so it
takes a while until the user gets the error back
and is bad UX. So we simplify this by checking
ourselves if we are logged in before submitting
the request.