ct/frontend: allow concurrent requests per producer
Currently, we can only upload a single batch from a producer at once in
order to preserve the ordering of requests into rm_stm for idempotency.

That is undesirable for both latency and cost (of s3 uploads). To fix
this, mark requests as enqueued before they are uploaded into s3, but
mark their order. Once the request is uploaded we re-enforce ordering
of batches to preserve the requirements for rm_stm to do its thing.