`tree`: Use `performance-move-const-arg.CheckMoveToConstRef` in `clang-tidy`
Oh dear, what a rabbit hole I have found myself in. What started as a bug fix
in an unrelated piece of code in which I accidentally `std::move()`'ed a
value into a co-routine that actually takes a `const T&` and ended up shooting
myself in the foot via stack-use-after-free, I realized that this issue
of `std::move()`'ing not actually `std::move()`'ing something is present
throughout the codebase. For a codebase rife with co-routines and lifetime
issues, this is a really bad thing.

So, let's enable this `clang-tidy` pass that should make us all a bit safer.